+++ client/tests/test_page_concordance.js
@@ -1,5 +1,5 @@
 "use strict";
-/*jslint plusplus: true */
+/*jslint plusplus: true, nomen: true */
 /*global Promise */
 var test = require('tape');
 var proxyquire =  require('proxyquire');
@@ -55,14 +55,14 @@ test('post_process', function (t) {
                     ['a', ' ', 'fine', ' ', 'mess', [0, 2, 4]],
                     ["you've", ' ', 'gotten', ' ', 'us', [0, 2, 4]],
                     ["into", '.', [0]],
-                    [['parp']], // Book ID
-                    [[1234]], // Word ID
+                    ['parp', 0, 100, 103], // Book / chapter / offset start / offset end
+                    [0, 0], // para-in-chap, sent-in-chap
                 ],
             ], chapter_start: {
                 parp: {0: 100},
             }, version: 33}
         );
-        t.deepEqual(data.data[0].DT_RowId, 'parp1234', 'Formed row ID from book/word');
+        t.deepEqual(data.data[0].DT_RowId, 'parp:0:100', 'Formed row ID from book/word');
         t.deepEqual(data.data[0].DT_RowClass, '', 'No matches, RowClass (explicitly) empty');
         t.deepEqual(data.data[0].kwic, 0, 'No matches, kwic 0');
         t.deepEqual(data.allWords, {}, 'Ignoring all rows, so no words found');
@@ -79,14 +79,14 @@ test('post_process', function (t) {
                     ['a', ' ', 'fine', ' ', 'mess', [0, 2, 4]],
                     ["you've", ' ', 'gotten', ' ', 'us', [0, 2, 4]],
                     ["into", '.', [0]],
-                    [['parp']], // Book ID
-                    [[1238]], // Word ID
+                    ['parp', 0, 120, 123], // Book / chapter / offset start / offset end
+                    [0, 0], // para-in-chap, sent-in-chap
                 ],
             ], chapter_start: {
                 parp: {0: 100},
             }, version: 33}
         );
-        t.deepEqual(data.data[0].DT_RowId, 'parp1238', 'Formed row ID from book/word');
+        t.deepEqual(data.data[0].DT_RowId, 'parp:0:120', 'Formed row ID from book/word');
         t.deepEqual(data.data[0].DT_RowClass, 'kwic-highlight-2', "Colour classes start at 2");
         t.deepEqual(data.data[0].kwic, 1, 'Overall 1 match');
         t.deepEqual(data.allWords, {'you\'ve': true, gotten: true, us: true, into: true}, 'Found words in middle, right');
@@ -106,26 +106,26 @@ test('post_process', function (t) {
                     ['a', ' ', 'fine', ' ', 'mess', [0, 2, 4]],
                     ["you've", ' ', 'gotten', ' ', 'us', [0, 2, 4]],
                     ["into", '.', [0]],
-                    ['parp', 99], // Book ID
-                    [1238, 5555], // Word ID
+                    ['parp', 99, 100, 103], // Book / chapter / offset start / offset end
+                    [0, 0], // para-in-chap, sent-in-chap
                 ],
                 [
                     ['a', ' ', 'fine', ' ', 'mess', [0, 2, 4]],
                     ["you've", ' ', 'gotten', ' ', 'out', [0, 2, 4]],
                     ["of", '.', [0]],
-                    ['slarp', 99], // Book ID
-                    [1238, 2222], // Word ID
+                    ['slarp', 99, 100, 103], // Book / chapter / offset start / offset end
+                    [0, 0], // para-in-chap, sent-in-chap
                 ],
                 [
                     ['a', ' ', 'fine', ' ', 'mess', [0, 2, 4]],
                     ["you've", ' ', 'gotten', ' ', 'away', [0, 2, 4]],
                     ["from", '.', [0]],
-                    ['parp', 99], // Book ID
-                    [1240, 5555], // Word ID
+                    ['parp', 99, 100, 103], // Book / chapter / offset start / offset end
+                    [0, 0], // para-in-chap, sent-in-chap
                 ],
             ], chapter_start: {
-                parp: {0: 100},
-                slarp: {0: 100},
+                parp: {0: 100, _end: 5555},
+                slarp: {0: 100, _end: 2222},
             }, version: 33}
         );