+++ server/clic/concordance.py
@@ -147,7 +147,7 @@ def create_concordance(cdb, q, result_set, contextsize, metadata):
     book_ids = set()
     for result in result_set:
         ch = cdb.get_chapter(result.id)
-        (book_id, chapter_num, count_prev_chap, total_word) = cdb.get_chapter_word_counts(result.id)
+        (book_id, chapter_num, _, _) = cdb.get_chapter_word_counts(result.id)
         book_ids.add(book_id)
 
         for match in result.proxInfo:
@@ -155,8 +155,8 @@ def create_concordance(cdb, q, result_set, contextsize, metadata):
             (word_id, para_chap, sent_chap) = cdb.get_word(result.id, match[0])
 
             conc_line = ch.get_conc_line(word_id, len(match), contextsize) + [
-                [book_id, chapter_num, para_chap, sent_chap],
-                [count_prev_chap + int(word_id), total_word, result.id, word_id, word_id + len(match)],
+                [book_id, chapter_num, word_id, word_id + len(match), result.id],
+                [para_chap, sent_chap],
             ]
 
             yield conc_line