+++ server/clic/subset.py
@@ -111,13 +111,13 @@ def subset(cdb, corpora=['dickens'], subset=['all'], contextsize=['0'], metadata
         if not cur_chapter or cur_chapter != chapter_id:
             cur_chapter = chapter_id
             ch = cdb.get_chapter(cur_chapter)
-            (book_id, chapter_num, count_prev_chap, total_word) = cdb.get_chapter_word_counts(chapter_id)
+            (book_id, chapter_num, _, _) = cdb.get_chapter_word_counts(chapter_id)
             book_ids.add(book_id)
 
         (_, para_chap, sent_chap) = cdb.get_word(chapter_id, [0, offset_start])
         yield ch.get_conc_line(offset_start, offset_end - offset_start, contextsize) + [
-                [book_id, chapter_num, para_chap, sent_chap],
-                [count_prev_chap + int(offset_start), total_word, chapter_id, offset_start, offset_end],
+                [book_id, chapter_num, offset_start, offset_end, chapter_id],
+                [para_chap, sent_chap],
         ]
 
     footer = cdb.get_book_metadata(book_ids, set(metadata))